Swap the soft link and the hard link
